handlers/scheduler.py
from telegram import Update
from telegram.ext import ContextTypes
from datetime import datetime
import pytz

async def schedule_message(update: Update, context: ContextTypes.DEFAULT_TYPE):
    # Implementação básica do agendamento
    await update.message.reply_text(
        "Para agendar uma mensagem, use o formato:\n"
        "/agendar HH:MM grupo_id mensagem\n"
        "Exemplo: /agendar 14:30 -1001234567890 Promoção ativa!"
    ) 